Converted test_api_logging.py
